op-program: Reflective program for block execution
