+++ op-e2e/system/proofs/system_fpp_test.go
@@ -298,7 +298,7 @@ func testFaultProofProgramScenario(t *testing.T, ctx context.Context, sys *e2esy
 	// Check the FPP confirms the expected output
 	t.Log("Running fault proof in fetching mode")
 	log := testlog.Logger(t, log.LevelInfo)
-	err := opp.FaultProofProgram(ctx, log, fppConfig)
+	err := opp.FaultProofProgramWithDefaultPrefecher(ctx, log, fppConfig)
 	require.NoError(t, err)
 
 	t.Log("Shutting down network")
@@ -315,13 +315,13 @@ func testFaultProofProgramScenario(t *testing.T, ctx context.Context, sys *e2esy
 	// Should be able to rerun in offline mode using the pre-fetched images
 	fppConfig.L1URL = ""
 	fppConfig.L2URL = ""
-	err = opp.FaultProofProgram(ctx, log, fppConfig)
+	err = opp.FaultProofProgramWithDefaultPrefecher(ctx, log, fppConfig)
 	require.NoError(t, err)
 
 	// Check that a fault is detected if we provide an incorrect claim
 	t.Log("Running fault proof with invalid claim")
 	fppConfig.L2Claim = common.Hash{0xaa}
-	err = opp.FaultProofProgram(ctx, log, fppConfig)
+	err = opp.FaultProofProgramWithDefaultPrefecher(ctx, log, fppConfig)
 	if s.Detached {
 		require.Error(t, err, "exit status 1")
 	} else {
