+++ op-program/host/common/l2_client.go
@@ -1,4 +1,4 @@
-package host
+package common
 
 import (
 	"context"
